+++ cogs/search.py
@@ -34,10 +34,11 @@ class SearchCog(commands.Cog, name="Search"):
             ctx {discord.ext.commands.Context} -- Context of the command.
             school {str} -- Name of the school the user wants to validate.
         """
-        await ctx.send(await utils.school_check(school))
+        embed = await utils.make_embed(ctx, title=str(await utils.school_check(school)))
+        await ctx.send(embed=embed)
 
     @commands.command(name="search-school",
-                      help="Search all schools for <school>.\nCollege, University, Community are blocked as they return a log of results.")  # noqa: E501 pylint: disable=line-too-long
+                      help="Search all schools for <school>.\nCollege, University, Community are blocked as they return a lot of results.")  # noqa: E501 pylint: disable=line-too-long
     async def search_school(self, ctx, *, school: str):
         """search-school
         ---
@@ -55,15 +56,17 @@ class SearchCog(commands.Cog, name="Search"):
                  school.lower() == "community"
                  ]
         if any(rules):
-            await ctx.send(
-                "Please refine your search as \"{}\" returns a lot of results ".format(school))
+            embed = await utils.make_embed(ctx, title="Search error",
+                                           description="Please refine your search as \"{}\" returns a lot of results ".format(school),  # noqa: E501 pylint: disable=line-too-long
+                                           color="FF0000"
+                                           )
+            await ctx.send(embed=embed)
             return
-        async with ctx.typing():
-            results = await utils.school_search(school)
-            if not results:
-                await ctx.send("No results found.")
-            else:
-                await utils.list_message(ctx, results, "Search Results:\n")
+        results = await utils.school_search(school)
+        if len(results) == 0:
+            await ctx.send("No results found.")
+        else:
+            await utils.list_message(ctx, results, "Search Results:\n")
 
     @commands.command(name="search-state")
     async def search_state(self, ctx, *, state: str):